+++ src/core/tools/common/native_aio_provider.posix.cpp
@@ -57,7 +57,7 @@ namespace dsn {
         error_code native_posix_aio_provider::close(dsn_handle_t hFile)
         {
             // TODO: handle failure
-            ::close(static_cast<int>(hFile));
+            ::close((int)(uintptr_t)(hFile));
             return ERR_OK;
         }
